hw2/requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asttokens==3.0.0
black==25.1.0
click==8.1.8
colorama==0.4.6
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.0
grpcio==1.71.0
icecream==2.1.4
Jinja2==3.1.6
lightning-utilities==0.14.1
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
protobuf==6.30.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyYAML==6.0.2
scipy==1.15.2
setuptools==76.0.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torchmetrics==1.6.3
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
Werkzeug==3.1.3
